refactor: move some extensions to widgets
